initial commit
This was originally a subfolder from my difforge repo that I'm now moving to be its own submodule. The code was getting too complicated, so I moved it out. Some things in here are depricated and no longer work as intended. I will remove them from the repo once they are tracked